Frontend Dev

+++ app.py
@@ -1,13 +1,15 @@
 from flask import Flask, render_template, request, redirect, make_response
-from ticket import Ticket, Constant, cookiePool, initChecker
+from ticket import Ticket, Constant, cookiePool, initChecker, InputValidator
 
 rootUsername = 'root'
 rootPassword = 'root'
 rootMailAddr = 'root@ticket.system'
 
+
 app = Flask(__name__)
 ticket = Ticket('./backend')
 ini = initChecker('inited.txt')
+validator = InputValidator()
 
 userPool = cookiePool()
 
@@ -30,8 +32,16 @@ def checkCookie():
 def preCheck():
     return ini.check() and checkCookie()
 
+@app.route('/down')
+def down():
+    if not ini.down():
+        return redirect('/')
+    return 'SERVER IS DOWN!'
+
 @app.route('/login', methods=("GET", "POST"))
 def login():
+    if ini.down():
+        return redirect('/down')
     if not ini.check():
         return redirect('/')
     if checkCookie():
@@ -41,6 +51,9 @@ def login():
         return render_template('form.html', flag = True, form = Constant.login_form, form_path = '/login', title = 'User Login')
 
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', flag = True, message = 'Illegal Input!', form = Constant.login_form, form_path = '/login', title = 'User Login')
+
     check = ticket.login(form['username'], form['password'])
     if not check:
         return render_template('form.html', flag = True, message = 'Failed', form = Constant.login_form, form_path = '/login', title = 'User Login')
@@ -52,6 +65,8 @@ def login():
 
 @app.route('/logout')
 def logout():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     userID = request.cookies.get('id')
@@ -75,9 +90,15 @@ def checkPrivilege(): # users with privilege level >= 1 are administrators.
 
 @app.route('/init', methods=("GET", "POST"))
 def init():
+    if ini.down():
+        return redirect('/down')
     if request.method == "GET":
         return render_template('form.html', flag = True, form = Constant.init_form, form_path = '/init', title = 'System Initialization')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', message = 'Illegal Input!', flag = True, form = Constant.init_form, form_path = '/init', title = 'System Initialization')
+
     if form['password'] != form['password2']:
         return render_template('form.html', message = 'Failed', flag = True, form = Constant.init_form, form_path = '/init', title = 'System Initialization')
 
@@ -91,6 +112,8 @@ def init():
 
 @app.route('/')
 def root():
+    if ini.down():
+        return redirect('/down')
     if not ini.check():
         return redirect('/init')
     userID = request.cookies.get('id')
@@ -102,11 +125,17 @@ def root():
 
 @app.route('/add_user', methods=("GET", "POST"))
 def add_user():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.add_user_form, form_path = '/add_user', title = 'Add User')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.add_user_form, form_path = '/add_user', title = 'Add User')
+
     ret = ticket.add_user(getUsername(), form['username'], form['password'], form['name'], form['mailAddr'], form['privilege'])
     if ret:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Succeed!', form = Constant.add_user_form, form_path = '/add_user', title = 'Add User')
@@ -115,11 +144,17 @@ def add_user():
 
 @app.route('/query_profile', methods=("GET", "POST"))
 def query_profile():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.query_profile_form, form_path = '/query_profile', title = 'Query Profile')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.query_profile_form, form_path = '/query_profile', title = 'Query Profile')
+
     ret = ticket.query_profile(getUsername(), form['username'])
     if ret == -1:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_profile_form, form_path = '/query_profile', title = 'Query Profile')
@@ -127,11 +162,17 @@ def query_profile():
 
 @app.route('/modify_profile', methods=("GET", "POST"))
 def modify_profile():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.modify_profile_form, form_path = '/modify_profile', title = 'Modify Profile')
+
     form = request.form.to_dict()
+    print('form got', form)
+    if not validator.check_modifyProfile(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.modify_profile_form, form_path = '/modify_profile', title = 'Modify Profile')
 
     if form['password'] != form['password2']:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.modify_profile_form, form_path = '/modify_profile', title = 'Modify Profile')
@@ -148,11 +189,17 @@ def modify_profile():
 
 @app.route('/add_train', methods=("GET", "POST"))
 def add_train():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.add_train_form, form_path = '/add_train', title = 'Add Train')
+
     form = request.form.to_dict()
+    if not validator.check_addTrain(form, len(form['stations'].split('|'))):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.add_train_form, form_path = '/add_train', title = 'Add Train')
+
     ret = ticket.add_train(form['trainID'], len(form['stations'].split('|')), form['seatNum'], form['stations'],
                            form['prices'], form['startTime'], form['travelTimes'], form['stopoverTimes'], form['saleDate'], form['type'])
     if ret:
@@ -162,11 +209,17 @@ def add_train():
 
 @app.route('/release_train', methods=("GET", "POST"))
 def release_train():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.release_train_form, form_path = '/release_train', title = 'Release Train')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.release_train_form, form_path = '/release_train', title = 'Release Train')
+
     ret = ticket.release_train(form['trainID'])
     if ret:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Succeed!', form = Constant.release_train_form, form_path = '/release_train', title = 'Release Train')
@@ -175,23 +228,35 @@ def release_train():
 
 @app.route('/query_train', methods=("GET", "POST"))
 def query_train():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.query_train_form, form_path = '/query_train', title = 'Query Train')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.query_train_form, form_path = '/query_train', title = 'Query Train')
+
     ret = ticket.query_train(form['trainID'], form['date'])
     if ret == -1:
-        render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_train_form, form_path = '/query_train', title = 'Query Train')
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_train_form, form_path = '/query_train', title = 'Query Train')
     return render_template('table.html', privilege = checkPrivilege(), username = getFriendlyname(), title = 'Train', ret = ret, ret_len = len(ret), table_head = Constant.query_train_table_head, col_list = Constant.query_train_list)
 
 @app.route('/delete_train', methods=("GET", "POST"))
 def delete_train():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.delete_train_form, form_path = '/delete_train', title = 'Delete Train')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.delete_train_form, form_path = '/delete_train', title = 'Delete Train')
+
     ret = ticket.delete_train(form['trainID'])
     if ret:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Succeed!', form = Constant.delete_train_form, form_path = '/delete_train', title = 'Delete Train')
@@ -200,23 +265,36 @@ def delete_train():
 
 @app.route('/query_ticket', methods=("GET", "POST"))
 def query_ticket():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.query_ticket_form, form_path = '/query_ticket', title = 'Query Ticket')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input', form = Constant.query_ticket_form, form_path = '/query_ticket', title = 'Query Ticket')
+
+
     ret = ticket.query_ticket(form['time'], form['start'], form['end'], form['sort_param'])
     if ret == -1:
-        render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_ticket_form, form_path = '/query_ticket', title = 'Query Ticket')
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_ticket_form, form_path = '/query_ticket', title = 'Query Ticket')
     return render_template('table.html', privilege = checkPrivilege(), username = getFriendlyname(), title = 'Ticket', ret = ret, ret_len = len(ret), table_head = Constant.query_ticket_table_head, col_list = Constant.query_ticket_list)
 
 @app.route('/query_transfer', methods=("GET", "POST"))
 def query_transfer():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.query_transfer_form, form_path = '/query_transfer', title = 'Query Transfer')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input!', form = Constant.query_transfer_form, form_path = '/query_ticket', title = 'Query Transfer')
+
     ret = ticket.query_transfer(form['time'], form['start'], form['end'], form['sort_param'])
     if ret == -1:
         render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', form = Constant.query_transfer_form, form_path = '/query_ticket', title = 'Query Transfer')
@@ -225,11 +303,17 @@ def query_transfer():
 
 @app.route('/buy_ticket', methods=("GET", "POST"))
 def buy_ticket():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.buy_ticket_form, form_path = '/buy_ticket', title = 'Buy Ticket')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegal Input', form = Constant.buy_ticket_form, form_path = '/buy_ticket', title = 'Buy Ticket')
+
     que = '1' if 'que' in form else '0'
     ret = ticket.buy_ticket(getUsername(), form['trainID'], form['day'], form['ffrom'], form['to'], form['number'], que)
     if ret == -1:
@@ -241,6 +325,8 @@ def buy_ticket():
 
 @app.route('/query_order')
 def query_order():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     ret = ticket.query_order(getUsername()) # it can never be -1
@@ -249,11 +335,17 @@ def query_order():
 
 @app.route('/refund_ticket', methods=("GET", "POST"))
 def refund_ticket():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), form = Constant.refund_ticket_form, form_path = '/refund_ticket', title = 'Refund Ticket')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Illegam Input!', form = Constant.refund_ticket_form, form_path = '/refund_ticket', title = 'Refund Ticket')
+
     ret = ticket.refund_ticket(getUsername(), form['num'])
     if ret:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Succeed!', form = Constant.refund_ticket_form, form_path = '/refund_ticket', title = 'Refund Ticket')
@@ -262,11 +354,15 @@ def refund_ticket():
 
 @app.route('/clean', methods=("GET", "POST"))
 def clean():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), warning = 'Are You Sure to Clear All Data?', form = Constant.clean_form, form_path = '/clean', title = 'Clear Data')
+
     form = request.form.to_dict()
+
     if form['username'] == getUsername():
         logoutRoot()
         ticket.clean()
@@ -279,6 +375,8 @@ def clean():
 
 @app.route('/shutdown', methods=("GET", "POST"))
 def shutdown():
+    if ini.down():
+        return redirect('/down')
     if not preCheck() or not checkPrivilege():
         return redirect('/')
     if request.method == "GET":
@@ -287,12 +385,15 @@ def shutdown():
     if form['username'] == getUsername():
         logoutRoot()
         ticket.exit()
+        ini.shutdown()
         return redirect('/')
     else:
         return render_template('form.html', privilege = checkPrivilege(), username = getFriendlyname(), message = 'Failed!', warning = 'Are You Sure to Shutdown the System?', form = Constant.shutdown_form, form_path = '/shutdown', title = 'Shutdown System')
 
 @app.route('/buy_ticket2', methods=("GET", "POST"))
 def buy_ticket2():
+    if ini.down():
+        return redirect('/down')
     if not preCheck():
         return redirect('/')
     if request.method == "GET": # unable to be accessed by get
@@ -302,6 +403,9 @@ def buy_ticket2():
     if not ('number' in form):
         return render_template('buy_ticket2.html', privilege = checkPrivilege(), username = getFriendlyname(), trainID = form['trainID'], day = form['day'], ffrom = form['ffrom'], to = form['to'])
     else:
+        if not validator.check_Normal(form):
+            return render_template('buy_ticket2.html', message = 'Illegal Input!', privilege = checkPrivilege(), username = getFriendlyname(), trainID = form['trainID'], day = form['day'], ffrom = form['ffrom'], to = form['to'])
+
         que = '1' if 'que' in form else '0'
         ret = ticket.buy_ticket(getUsername(), form['trainID'], form['day'], form['ffrom'], form['to'], form['number'], que)
         if ret == -1:
@@ -313,11 +417,17 @@ def buy_ticket2():
 
 @app.route('/register', methods=("GET", "POST"))
 def register():
+    if ini.down():
+        return redirect('/down')
     if not ini.check():
         return redirect('/')
     if request.method == "GET":
         return render_template('form.html', flag = True, form = Constant.register_form, form_path = '/register', title = 'User Registration')
+
     form = request.form.to_dict()
+    if not validator.check_Normal(form):
+        return render_template('form.html', flag = True, message = 'Illegal Input!', form = Constant.register_form, form_path = '/register', title = 'User Registration')
+
     if form['password'] != form['password2']:
         return render_template('form.html', flag = True, message = 'Failed', form = Constant.register_form, form_path = '/register', title = 'User Registration')
 
@@ -327,6 +437,15 @@ def register():
         return redirect('/login')
     else:
         return render_template('form.html', flag = True, message = 'Failed', form = Constant.register_form, form_path = '/register', title = 'User Registration')
+
+@app.route('/user_profile')
+def getProfile():
+    if ini.down():
+        return redirect('/down')
+    if not preCheck():
+        return redirect('/')
+    ret = ticket.query_profile(getUsername(), getUsername())
+    return render_template('query_profile.html', privilege = checkPrivilege(), username = getFriendlyname(), ret = ret)
 
 initRoot()
